requirements.txt
asttokens==2.4.1
backcall==0.2.0
basketball-reference-web-scraper==4.15.2
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
fonttools==4.55.0
gower==0.1.2
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.7
llvmlite==0.41.1
lxml==5.3.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.58.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
POT==0.9.5
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
